add file name lint (#2417)

+++ Makefile.d/functions.mk
@@ -337,7 +337,14 @@ define update-github-actions
 		if [ -n "$$ACTION_NAME" ]; then \
 			FILE_NAME=`echo $$ACTION_NAME | tr '/' '_' | tr '-' '_' | tr '[:lower:]' '[:upper:]'`; \
 			if [ -n "$$FILE_NAME" ]; then \
-				VERSION=`curl --silent https://api.github.com/repos/$$ACTION_NAME/releases/latest | grep -Po '"tag_name": "\K.*?(?=")' | sed 's/v//g'`;\
+				if [ "$$ACTION_NAME" = "aquasecurity/trivy-action" ] || [ "$$ACTION_NAME" = "machine-learning-apps/actions-chatops" ]; then \
+					VERSION="master"; \
+				elif [ "$$ACTION_NAME" = "softprops/action-gh-release" ]; then \
+					VERSION="1.0.0"; \
+				else \
+					REPO_NAME=`echo $$ACTION_NAME | cut -d'/' -f1-2`; \
+					VERSION=`curl --silent https://api.github.com/repos/$$REPO_NAME/releases/latest | grep -Po '"tag_name": "\K.*?(?=")' | sed 's/v//g' | sed -E 's/[^0-9.]+//g'`;\
+				fi; \
 				if [ -n "$$VERSION" ]; then \
 					echo "updating $$ACTION_NAME version file $$FILE_NAME to $$VERSION"; \
 					echo $$VERSION > $(ROOTDIR)/versions/actions/$$FILE_NAME; \
@@ -345,8 +352,16 @@ define update-github-actions
 					VERSION=`cat $(ROOTDIR)/versions/actions/$$FILE_NAME`; \
 					echo "No version found for $$ACTION_NAME version file $$FILE_NAME=$$VERSION"; \
 				fi; \
-				VERSION_PREFIX=`echo $$VERSION | cut -c 1`; \
-				find $(ROOTDIR)/.github -type f -exec sed -i "s%$$ACTION_NAME@.*%$$ACTION_NAME@v$$VERSION_PREFIX%g" {} +; \
+				if [ "$$ACTION_NAME" = "cirrus-actions/rebase" ]; then \
+					VERSION_PREFIX=$$VERSION; \
+					find $(ROOTDIR)/.github -type f -exec sed -i "s%$$ACTION_NAME@.*%$$ACTION_NAME@$$VERSION_PREFIX%g" {} +; \
+				elif echo $$VERSION | grep -qE '^[0-9]'; then \
+					VERSION_PREFIX=`echo $$VERSION | cut -c 1`; \
+					find $(ROOTDIR)/.github -type f -exec sed -i "s%$$ACTION_NAME@.*%$$ACTION_NAME@v$$VERSION_PREFIX%g" {} +; \
+				else \
+					VERSION_PREFIX=$$VERSION; \
+					find $(ROOTDIR)/.github -type f -exec sed -i "s%$$ACTION_NAME@.*%$$ACTION_NAME@$$VERSION_PREFIX%g" {} +; \
+				fi; \
 			else \
 				echo "No action version file found for $$ACTION_NAME version file $$FILE_NAME" >&2; \
 			fi \
@@ -355,3 +370,4 @@ define update-github-actions
 		fi \
 	done
 endef
+
